did some generalsort

+++ src/experiment/InsertionSort.java
@@ -3,12 +3,12 @@ package experiment;
 import java.util.Comparator;
 
 public class InsertionSort {
-    public static <E extends Comparable<E>> int sort(E[] array) {
+    public static <E extends Comparable<E>> long sort(E[] array) {
         return sort(array, (x, y) -> x.compareTo(y));
     }
 
-    public static <E> int sort(E[] array, Comparator<E> comparator) {
-        int steps = 0;
+    public static <E> long sort(E[] array, Comparator<E> comparator) {
+        long steps = 0;
         for (int i = 1; i < array.length; i++) {
             E x = array[i];
             int j;
